VMTest with atomic op tests split to several smaller files
